Use set_networks_for_node esi l2 task
Ensure that a only a specific set of networks is attached to a node.
Previously, we only bothered attaching new networks, but didn't bother to
detach existing network connections. This could lead to unexpected
behavior.

The set_networks_for_node tasks ensures that only the networks we intend
are attached to a node.