Get worker node address information from ESI
We were previously getting the address of a worker node (to be used as the
target for the ingress port forwarding) from the agent inventory, but this
information can be inaccurate if an agent has just been moved to a
different network. This would cause the port forwarding to fail.

With this commit, we get the node address from ESI.