Don't require manage_agents_cluster_network
When deleting a cluster, we still call the
`attach_and_approve_all_new_agents` task list, which requires the
manage_agents_cluster_network -- which *indirectly* requires
`create_network_result`. We need to refactor where we're setting the
network name, but for now we can just mark this parameter as not required.