Use the private API to publish templates
Now that the multi-tenancy logic is in place templates need to be
published using the private API, as otherwise they will be assigned to
the tenant of the user publishing them, and won't be usable by other
users. Using the private API ensures that they are assigned to the
`shared` tenant, which is visible for everyone.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>